# src/main/resources/static/swagger.yaml
swagger: '2.0'
info:
  description: This the InnoEdu api.
  version: 1.0.0
  title: InnoEdu Api
  contact:
    email: apiteam@inno.edu
host: inno-edu-api.herokuapp.com
basePath: /api
...
responses:
  400:
    description: Bad request, additional info provided in the response body
    schema:
      $ref: '#/definitions/VndErrors'
...
  403:
    description: Forbidden, access is denied
...
securityDefinitions:
  forwarded:
    name: authorization
    type: apiKey
    in: header
schemes:
  - https
  - http
paths:
...
  /profiles:
    get:
      tags:
        - profiles
      summary: Get all profiles
      produces:
        - application/hal+json
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/EmbeddedProfileResourceList'
        '403':
          $ref: '#/responses/403'
    post:
      tags:
        - profiles
      summary: Add a new profile
      consumes:
        - application/json
      produces:
        - application/hal+json
      parameters:
        - in: body
          name: request
          description: Request object with the profile information to be created
          required: true
          schema:
            $ref: '#/definitions/CreateProfileRequest'
      responses:
        '201':
          description: New profile successfully created
          schema:
            $ref: '#/definitions/Profile'
          headers:
            Location:
              type: string
              description: Link to the new resource created
        '400':
          $ref: '#/responses/400'
        '403':
          $ref: '#/responses/403'
  /profiles/{id}:
    get:
      tags:
        - profiles
      summary: Get profile by id
      parameters:
        - in: path
          name: id
          description: Profile id
          required: true
          type: string
          format: uuid
      produces:
        - application/hal+json
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/ProfileResource'
        '404':
          description: Profile not found
          schema:
            $ref: '#/definitions/VndErrors'
        '403':
          $ref: '#/responses/403'
    put:
      tags:
        - profiles
      summary: Update profile by id
      consumes:
        - application/json
      produces:
        - application/hal+json
      parameters:
        - in: path
          name: id
          description: Profile id
          required: true
          type: string
          format: uuid
        - in: body
          name: request
          description: Request object with the profile information to be updated
          required: true
          schema:
            $ref: '#/definitions/UpdateProfileRequest'
      responses:
        '201':
          description: Profile successfully updated
          schema:
            $ref: '#/definitions/Profile'
        '404':
          description: Profile not found
          schema:
            $ref: '#/definitions/VndErrors'
        '400':
          $ref: '#/responses/400'
        '403':
          $ref: '#/responses/403'
    delete:
      tags:
        - profiles
      summary: Delete profile by id
      parameters:
        - in: path
          name: id
          description: Profile id
          required: true
          type: string
          format: uuid
      responses:
        '204':
          description: Profile successfully removed
        '404':
          description: Profile not found
          schema:
            $ref: '#/definitions/VndErrors'
        '403':
          $ref: '#/responses/403'
  /profiles/{id}/associations:
    get:
      tags:
        - profile-associations
      summary: Get profile associations by profile and status
      parameters:
        - in: path
          name: id
          description: Profile id
          required: true
          type: string
          format: uuid
        - in: query
          name: status
          description: Status
          required: false
          type: string
          enum: [PENDING, APPROVED, REJECTED]
      produces:
        - application/hal+json
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/ProfileAssociationResourceList'
        '404':
          description: Profile not found
          schema:
            $ref: '#/definitions/VndErrors'
        '403':
          $ref: '#/responses/403'
  /profiles/{id}/associate:
    post:
      tags:
        - profile-associations
      summary: Associate a profile to a school
      consumes:
        - application/json
      produces:
        - application/hal+json
      parameters:
        - in: path
          name: id
          description: Profile id
          required: true
          type: string
          format: uuid
        - in: body
          name: request
          description: Request object with the school information to be associated
          required: true
          schema:
            $ref: '#/definitions/ProfileAssociationRequest'
      responses:
        '200':
          description: Profile successfully associated
          schema:
            $ref: '#/definitions/Profile'
        '404':
          description: Profile not found
          schema:
            $ref: '#/definitions/VndErrors'
        '400':
          $ref: '#/responses/400'
        '403':
          $ref: '#/responses/403'
...
  /availability/profile/{profileId}:
    get:
      tags:
        - availability
      summary: Get availability for a profile
      parameters:
        - in: path
          name: profileId
          description: Profile id
          required: true
          type: string
          format: uuid
      produces:
        - application/hal+json
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/EmbeddedAvailabilityResourceList'
        '403':
          $ref: '#/responses/403'
...
definitions:
  VndError:
    type: object
    properties:
      logref:
        type: string
      message:
        type: string
      links:
        type: array
        items:
          type: object
  VndErrors:
    type: array
    items:
      $ref: '#/definitions/VndError'
...
  CreateProfileRequest:
    type: object
    properties:
      userId:
        type: string
        format: uuid
      description:
        type: string
    required:
      - userId
      - description
  UpdateProfileRequest:
    type: object
    properties:
      description:
        type: string
    required:
      - description
  Profile:
    type: object
    properties:
      id:
        type: string
        format: uuid
      userId:
        type: string
        format: uuid
      schoolId:
        type: string
        format: uuid
      description:
        type: string
      rate:
        type: number
      status:
        type: string
        enum: [CREATED, ACTIVE, INACTIVE, REJECTED]
  ProfileResource:
      allOf:
        - $ref: '#/definitions/Profile'
        - type: object
          properties:
            _links:
              type: object
  ProfileResourceList:
    type: object
    properties:
      profileResourceList:
        type: array
        items:
          $ref: '#/definitions/ProfileResource'
  EmbeddedProfileResourceList:
    type: object
    properties:
      _embedded:
        $ref: '#/definitions/ProfileResourceList'
      _links:
        type: object
...
  Availability:
    type: object
    properties:
      id:
        type: string
        format: uuid
      mentorProfileId:
        type: string
        format: uuid
      fromDateTime:
        type: string
        format: date-time
      toDateTime:
        type: string
        format: date-time
  AvailabilityResource:
    allOf:
        - $ref: '#/definitions/Availability'
        - type: object
          properties:
            _links:
              type: object
  AvailabilityResourceList:
    type: object
    properties:
      availabilityResourceList:
        type: array
        items:
          $ref: '#/definitions/AvailabilityResource'
  EmbeddedAvailabilityResourceList:
    type: object
    properties:
      _embedded:
        $ref: '#/definitions/AvailabilityResourceList'
      _links:
        type: object
...
  ProfileAssociationRequest:
    type: object
    properties:
      schoolId:
        type: string
    required:
      - schoolId
  ProfileAssociation:
    type: object
    properties:
      id:
        type: string
        format: uuid
      profileId:
        type: string
        format: uuid
      schoolId:
        type: string
        format: uuid
      status:
        type: string
        enum: [PENDING, APPROVED, REJECTED]
      description:
        type: string
  ProfileAssociationResource:
    allOf:
        - $ref: '#/definitions/ProfileAssociation'
        - type: object
          properties:
            _links:
              type: object
  ProfileAssociationResourceList:
    type: object
    properties:
      profileAssociationResourceList:
        type: array
        items:
          $ref: '#/definitions/ProfileAssociationResource'